+++ pybop/costs/error_measures.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pybamm
 
 from pybop.costs.base_cost import BaseCost
 from pybop.costs.evaluation import Evaluation
@@ -71,10 +70,10 @@ class ErrorMeasure(BaseCost):
         else:
             self.weighting = np.asarray(weighting)
 
-    def evaluate(
+    def evaluate_batch(
         self,
-        solution: Solution | pybamm.Solution | FailedSolution,
-        inputs: Inputs | None = None,
+        solution: list[Solution],
+        inputs: list[Inputs],
         calculate_sensitivities: bool = False,
     ) -> Evaluation:
         """
@@ -82,38 +81,48 @@ class ErrorMeasure(BaseCost):
 
         Parameters
         ----------
-        solution : pybop.Solution | pybamm.Solution
-            The simulation result.
-        inputs : Inputs, optional
-            Input parameters (default: None).
+        Parameters
+        ----------
+        solution : list[Solution]
+            A list of simulation results.
+        inputs : list[Inputs]
+            The corresponding list of input parameters.
         calculate_sensitivities : bool
             Whether to also return the sensitivities (default: False).
         """
-        # Return failure cost if the solution failed
-        if isinstance(solution, FailedSolution):
-            return self.failure(self.parameters.names, calculate_sensitivities)
-
-        if not isinstance(solution, (Solution, pybamm.Solution)):
-            raise ValueError(
-                f"solution must be a pybop.Solution object, got type {type(solution)} with value {solution}."
-            )
-
-        # Early return if the prediction is not verified
-        if not self.verify_prediction(solution):
-            return self.failure(self.parameters.names, calculate_sensitivities)
-
-        # Compute the residual for all output variables
-        r = np.asarray(
-            [solution[var].data - self._target_data[var] for var in self.target]
+        e = np.empty(len(solution))
+        de = (
+            {key: np.zeros(len(solution)) for key in inputs[0].keys()}
+            if calculate_sensitivities
+            else None
         )
 
-        if calculate_sensitivities:
-            # Extract the sensitivities for all output variables and parameters
-            dy = self.stack_sensitivities(solution)
-            e, de = self.__call__(r=r, dy=dy, inputs=inputs)
-            return Evaluation(values=e, sensitivities=de)
+        for i, (sol, x) in enumerate(zip(solution, inputs, strict=False)):
+            # Return failure cost if the solution failed
+            if isinstance(sol, FailedSolution) or not self.verify_prediction(sol):
+                if calculate_sensitivities:
+                    e[i], de_i = self.failure(x.keys(), calculate_sensitivities)
+                else:
+                    e[i] = self.failure(x.keys(), calculate_sensitivities)
 
-        return Evaluation(values=self.__call__(r=r, inputs=inputs))
+            else:
+                # Compute the residual for all output variables
+                r = np.asarray(
+                    [sol[var].data - self._target_data[var] for var in self.target]
+                )
+
+                if calculate_sensitivities:
+                    # Extract the sensitivities for all output variables and parameters
+                    dy = self.stack_sensitivities(sol)
+                    e[i], de_i = self.__call__(r=r, dy=dy, inputs=x)
+                else:
+                    e[i] = self.__call__(r=r, inputs=x)
+
+            if calculate_sensitivities:
+                for key in x.keys():
+                    de[key][i] = de_i[key]
+
+        return Evaluation(values=e, sensitivities=de)
 
     def verify_prediction(self, solution: Solution):
         """